requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
mcp>=1.0.0
janome>=0.5.0
pymupdf>=1.23.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.21
chromadb>=0.4.0
requests>=2.31.0
boto3>=1.28.0
litellm>=1.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0

